+++ generate_pigeon.sh
@@ -1,4 +1,6 @@
 #!/bin/bash
-mkdir -p ios/AppMetricaPlugin/
+
+set -e
+
 mkdir -p android/src/main/java/io/appmetrica/analytics/flutter/pigeon
 dart run pigeon --input pigeons/appmetrica_api.dart